backend/requirements.txt
django==2.2.7
djangorestframework==3.10.3
drf-yasg==1.17.0
djangorestframework-jwt==1.11.0
django-phonenumber-field==4.0.0
django-nocaptcha-recaptcha==0.0.20
